requirements.txt
google-auth
GoogleAppEngineCloudStorageClient
GoogleAppEngineMapReduce
markdown2
python-jose
requests
requests_toolbelt
urllib3
